Consistent post media sizes (#219)
This PR aims to improve some small UX aspects around "endless scrolling"
while `PostMedia` loads images.

Previously, images within posts would size up to fit the width of the
list, with an unconstrained height. This led to cases where images up in
the list could push content down if and when they load late. A typical
real-world scenario could include a slower than normal internet
connection; if a user is reading some content below where an image would
normally be, and this image loads late, their "reading position" will be
snapped downward by the loaded image.

These changes include setting a constant size for images, loaded or
unloaded, to be 300dp. Once an image loads in, it is clipped and fit to
be contained within this boundary. Tapping on the image, or otherwise
viewing the entire post in full, will display the entire image
unclipped.

The following video illustrates the clipped image, as it transitions as
a hero to the full screen preview.


https://github.com/liftoff-app/liftoff/assets/459941/2a9220cb-b8cd-4301-81ed-e83c42827822

The following video illustrates the constant height while an image is
loading.


https://github.com/liftoff-app/liftoff/assets/459941/01895929-9a73-44cb-ac69-f6e2e3b464e2

Co-authored-by: swmarks <swmarks@users.noreply.github.com>