Auto-hide "not logged in" snackbar (#250)
Auto-hides the "You're not logged in, log in" snackbar after 7 seconds.

This Snackbar is displayed attempting to interact (post a comment,
upvote, downvote, etc) with an instance you're not logged into. The
current behavior is to display it indefinitely, and the Snackbar never
disappears on its own, you MUST click "log in" to make it disappear.