Pass username and instanceHost in UserData / pass UserData instead of JWT
The app mostly just passes the JWT token around, which is ok for auth
but is not ok for identification. This modifies a lot of files but all
it really does is refactors the controllers to use UserData instead of
just the JWT token directly. This will allow us to identify which user
the request is being made for, rather than just blindly passing auth
data. This will be more useful in later commits.