Add Chrome Custom Tabs/Safari View Controller (replace internal browser) (#238)
I didn't love the in-app view we had been using, so I added Chrome
Custom Tabs for Android and Safari View Controller for iOS. This should
maintain compatibility across OS's through conditional statements, but
testing is welcome.

I also changed the nomenclature for the settings page to be "Use
external browser:" and flipped the logic because they're both external
in a way.