Menu item to view a community in a specific instance other than the originating instance (#254)
This adds a new menu item which appears under the "..." menu of a
community called "View on another instance".

![image](https://github.com/liftoff-app/liftoff/assets/1340627/e7bb7f1a-fd88-444a-b823-1bb5289cb3b8)

This pops up a list of accounts for you to choose from. The "Subscribe"
button, when viewing anonymously, also pops this up:

![image](https://github.com/liftoff-app/liftoff/assets/1340627/3a2e257a-d9ac-407f-86df-e4cd3493e0fb)

This then executes a ResolveObject call to retrieve the community:

![image](https://github.com/liftoff-app/liftoff/assets/1340627/9ccd7b96-0aca-4f06-8b04-f8d1b5e69bf9)

Before showing you the community, in the other instance. Denoted by the
"(via instance)" text in the name of the community.

![image](https://github.com/liftoff-app/liftoff/assets/1340627/0c8c01d0-0cfd-47cd-8633-29f49ee31ece)

You can then subscribe directly from here, or if some posts are loaded
you can even comment or vote directly.

Dependent on #261