Revert "Add scroll chainer for post body (#286)" (#294)
This reverts commit af79d28e8216f6286f935bf02bc1cce12dbc9251.

This is causing all sorts of scrolling issues on the main post feed and
throughout the app and it's like the lists are getting locked in
position and you can't scroll all the way to the top. Unless we can fix
things easily as they are presently, I think we should revert this
commit/PR for the moment.