Revert the Image Preview changes because it needs to be done properly and with forethought (#290)
1. Scroll is still janky even with these image preview changes
2. While the scaled option is pretty to look at, it has some downsides
when used for browsing lemmy such as:

- > It can make the contents of images really small and hard to read,
requiring users to tap open the image anyways which is a waste of user's
time
- > There is quite literally an uproar and people want full height
images back: https://lemm.ee/post/701782
- > The post preview will also render images in the post body which can
cause issues with mismatched images sizes in the feed:

![image](https://github.com/liftoff-app/liftoff/assets/6200670/77712636-0a75-45e1-9003-11ce88896930)

If we're gonna do this differently we need to do it properly and with
forethought. We're also building on top of a solution that didnt
actually fix the scroll issue and just made users mad.