Add ability to block community from post menu (#206)
Fixes #32 

A couple of things to note. I didn't add in the `unblock/block` logic
that user blocking has as the fact that a community is blocked is not
readily available from the present records. I didn't want to add another
call just to check the community status. Getting to a post is also
pretty hard if you have blocked a community. Even your comments
disappear from your profile when you have blocked the whole community.

I wasn't able to figure out how to trigger a refresh of the infinite
list if someone blocks a community. The posts will stick around until
they refresh. This is also the current behaviour when you unblock a
community from the settings. You still need to manually refresh the list
to start seeing the posts again.


Cheers!