Support for video providers (#363)
This PR fixes #333 and #332

NOTE: This is using the temp creds for redgif and refreshing them on
every call. Eventually we should get real keys. There is an active
request out to redgif for our own keys.

The only difference between redgifs and mp4 playback is that we need to
request a url to use when playing back the redgif
Videos are set to automatically play, mute, and loop. I have not
implemented FullScreen support, but I figured even the thumbnails were a
good start.

I tested virtually using android and ios. The issue with android was
that the ExoPlayer would not accept a user-agent override.