Address linting issues after analysis options changes.
This includes a workaround for mobx stores:
https://github.com/mobxjs/mobx.dart/issues/809

Also, while uuid was already imported in fullscreenable_image.dart, it was not in pubspec.yaml.

Lastly, there are some linting issues with use_build_context_synchronously. For the files most heavily impacted, I have added a temporary ignore rule as the linting fixes are not currently in the Flutter SDK. See: https://github.com/dart-lang/linter/issues/4007