Fix state issues for posts (#425)
This depends on #424 .

Fixes #399 
Fixes #387
but we probably need a separate issue tracking the same thing for
comments.

This issue was that we store a list of `PostViews` in the
`InfinitePostList`, but then create a `PostStore`s with a reference to
those `PostView`s in `PostTile`. When updates happen (upvote, save,
etc.) we update the reference to the `PostView` in the `PostStore`, but
that doesn't change the original `PostView` stored in the
`InfinitePostList`. When the `PostTile` gets re-created due to scrolling
it creates a new `PostStore` with the old `PostView` from the parent and
the updated state is lost.

This fixes it by changing `InfinitePostList` to store a list of
`PostStore`s directly. This way, the reference doesn't get duplicated
and no state is lost when re-creating.

---------

Co-authored-by: Zach Russell <zachatrocity@gmail.com>