Add Flutter lints (#429)
Use flutter_lints package for recommended linting rules.

See #414 and #426.

My previous PR had an issue with the more complex mobx stores. I have
now reverted the related changes to those files and added an ignore rule
for each. This can be removed when Mobx is updated to be compliant with
flutter_lints.