Disable rounded corners/shadow and darken background of light theme (#442)
IMHO it looks much better. (I kind of stole the new background color
from a preview screenshot of Sync for Lemmy).

This also renames the options for rounding/shadow so that it resets
existing user settings.
That way, people have to see the new look, dislike it and _actively
choose_ to go back to the old look.


![Screenshot_20230723-005955.png](https://github.com/liftoff-app/liftoff/assets/856493/b9282522-811b-4e63-b995-7aaf8923f85d)
![Screenshot_20230723-005922.png](https://github.com/liftoff-app/liftoff/assets/856493/28149a48-7291-4406-9ed6-631f929bd220)