Fix CI by using fixed flutter version 3.10.6 (#518)
Building the pull requests by the CI stopped working some time ago.
Presumably, the specified flutter channel "stable" got updated and is
now incompatible which lets every build fail. Pinning the flutter
version to 3.10.6 for the CI fixes that. Note that the version was
chosen by best guess that worked at least somewhat locally. (I have yet
to get a build finished locally.) It might not be the version that was
used "originally".